Tests have been refactored with extending a new homemade WebTestCase class
