+++ web/test-coverage/Entity/Task.php.html
@@ -556,7 +556,7 @@
      <span class="warning"><strong>Dead Code</strong></span>
     </p>
     <p>
-     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 5.3.0</a> using <a href="https://secure.php.net/" target="_top">PHP 7.1.9</a> with <a href="https://xdebug.org/">Xdebug 2.6.0rc1</a> and <a href="https://phpunit.de/">PHPUnit 6.5.5</a> at Sat Jan 27 16:43:33 UTC 2018.</small>
+     <small>Generated by <a href="https://github.com/sebastianbergmann/php-code-coverage" target="_top">php-code-coverage 5.3.0</a> using <a href="https://secure.php.net/" target="_top">PHP 7.1.9</a> with <a href="https://xdebug.org/">Xdebug 2.6.0rc1</a> and <a href="https://phpunit.de/">PHPUnit 6.5.5</a> at Sat Jan 27 17:14:00 UTC 2018.</small>
     </p>
     <a title="Back to the top" id="toplink" href="#"><span class="glyphicon glyphicon-arrow-up"></span></a>
    </footer>
